6ca33d09c65d4bcebe847d32ad3f8c12/60-9-hamming-distance/src/test/java/HammingDistanceTest.java

import static org.junit.Assert.*;

import org.junit.Test;

public class HammingDistanceTest {

	@Test
	public void testZeroDistance() {
		int actualDistance = HammingDistance.hammingDistance(2, 2);
		assertEquals(0, actualDistance);
	}
	
	@Test
	public void testPositive() {
		int actualDistance = HammingDistance.hammingDistance(2, 4);
		assertEquals(2, actualDistance);
	}
	
	@Test
	public void testNegative() {
		int actualDistance = HammingDistance.hammingDistance(-2, -4);
		assertEquals(1, actualDistance);
	}
	
	@Test
	public void testPostiveNegative() {
		int actualDistance = HammingDistance.hammingDistance(2, -4);
		assertEquals(31, actualDistance);
	}
	
	@Test
	public void testZero() {
		int actualDistance = HammingDistance.hammingDistance(0, 0);
		assertEquals(0, actualDistance);
	}

}
